Provide original total prop to checkout-summary extension point (#324)
#### What problem is this solving?

The "Discounts" totalizer in the orderForm does not take into account
the difference between an item's list price and selling price. The
"Subtotal" shown in the minicart, therefore, is based on selling price
and the "Discounts" totalizer only appears if promotional discounts have
been applied.

This PR makes the "originalTotal" available (i.e. the sum of the items'
list prices multiplied by their quantity) as a prop passed to the
"checkout-summary" extension point. A separate PR will be opened on the
Checkout Summary app to allow this information to be rendered (if a
client opts in to doing so). This will allow a client to display
something like the following in their minicart:


![image](https://user-images.githubusercontent.com/6306265/210881636-da482086-ee5d-411c-97e3-c12f5554db4a.png)

#### How to test it?

Linked here:
https://minicartsubtotal--beautycounterqa.myvtex.com/nourishing-eye-cream/p